ILS, new constructive and refactoring.

+++ tmp/Solution.jl
@@ -2,7 +2,7 @@ mutable struct Solution
     routes::Vector{Vector{Vector{Int64}}}
 
     cost::Float64
-    route_cost::Float64
+    route_cost::Int64
     inventory_cost::Float64
 
     in_period::Vector{BitVector}
